% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addURL.R
\name{addURL}
\alias{addURL}
\title{Add a URL}
\usage{
addURL(url, text = NULL, extra = NULL, markdown = FALSE)
}
\arguments{
\item{url}{a URL.}

\item{text}{a character string to be searched within.}

\item{extra}{a connection object or a character string that stand for
a path, ignored if text is defined.}

\item{markdown}{a logical. If \code{TRUE}, hyperlinks marks of Markdown language are sought.}
}
\description{
Add a URL in HTML or Markdown format.
}
\examples{
addURL('https://github.com/inSileco/letiRmisc', 'Github repo')
addURL('https://github.com/inSileco/letiRmisc', 'Github repo', markdown=TRUE)
}
