% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addWebIcon.R
\name{addWebIcon}
\alias{addWebIcon}
\title{Add a web icon.}
\usage{
addWebIcon(name, framework = "fa", size = 2)
}
\arguments{
\item{name}{name of the icon to be added.}

\item{framework}{a character use to default is \code{fa} (which stands for Font-Awesome).}

\item{size}{default sets to \code{2},}
}
\description{
Generates a HTML tag to include an icon from an iconic framework.
}
\examples{
addWebIcon('cc')
}
\author{
Kevin Cazelles
}
