% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustString.R
\name{adjustString}
\alias{adjustString}
\title{Adjust the size of a character string to a given number of characters}
\usage{
adjustString(x, szchar, extra = 0, before = TRUE)
}
\arguments{
\item{x}{a character vector, or a vector to be coerced to a character vector.}

\item{szchar}{an integer indicating the size of the character string
to be created.}

\item{extra}{character vector, or a vector to be coerced to a character
vector that will be (partially) added to produced a string of}

\item{before}{logical. If \code{TRUE}, then the extra characters are added before
\code{x}}
}
\value{
A character vector of the concatenated characters.
}
\description{
Adjust the size of a character string to a given number of characters
by truncating it or adding specified extra characters.
}
\details{
This function was originally created to help getting a fixed number of digits
when naming files. The current version is more general but it remains
extremely useful to name files.
}
\examples{
paste0('myfilename', adjustString(c(1:2,10,100), 3))
adjustString('# Comment', 20, '#', before=FALSE)
}
\author{
Kevin Cazelles
}
