% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustStrings.R
\name{adjustStrings}
\alias{adjustStrings}
\title{Adjust the size of a vector of character strings}
\usage{
adjustStrings(x, n, extra = 0, before = TRUE)
}
\arguments{
\item{x}{a character vector, or a vector to be coerced to a character vector.}

\item{n}{a positive integer integer indicating the size of the
character string to be created.}

\item{extra}{a character vector, or a vector to be coerced to a character
vector that will be (partially) added to produced a string of a specific length.}

\item{before}{a logical. If \code{TRUE}, then the extra characters are added
before \code{x}.}
}
\value{
A character vector of the concatenated characters.
}
\description{
Adjust the size of a vector of character strings to a given number of
characters by truncating it or adding specified extra characters.
}
\details{
This function was originally created to help getting a fixed number of digits
when naming files. The current version is more general, it allows to
add any string before or after to adjust the size. Not that if a character
is longer than expected, it will be adequately cut off.
}
\examples{
paste0('myfilename', adjustStrings(c(1:2,10,100), 3))
adjustStrings('# Comment', 20, '#', before=FALSE)
}
\seealso{
\code{\link[base:sprintf]{base::sprintf()}}
}
