% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateCol.R
\name{aggregateCol}
\alias{aggregateCol}
\title{Aggregate columns}
\usage{
aggregateCol(data, grp, names_aggreg = NULL, FUN = sum, ...)
}
\arguments{
\item{data}{a data frame.}

\item{grp}{an integer vector, one value per \code{data} column. Integers are used
to group \code{data} columns, column with the same integer will be pooled
together.  Any integer is a valid group, \code{0} indicates that the column must
be unchanged, \code{NA} removes columns.}

\item{names_aggreg}{column names for aggregated columns.}

\item{FUN}{a function to be applied to all groups of columns.}

\item{...}{further arguments to be passed to \code{FUN}.}
}
\value{
A data frame with the grouped columns.
}
\description{
\code{aggregateCol()} performs summary statistics on a specific subset of columns.
}
\details{
\code{grp} is an integer vector, whose length is the number of columns of the
input data frame, is used to assign a group to each column that explicit the
sub-setting. Furthermore columns can be either kept as is using \code{0} or
discarded using \code{NA}.
}
\examples{
mat1 <- matrix(1:70,10)
grp1 <- c(NA,0, 1, 1, 2, 2, 2)
aggregateCol(mat1, grp1, FUN = mean)
}
\seealso{
\code{\link[stats:aggregate]{stats::aggregate()}} \link[base:split]{base::split}
}
