% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyString.R
\name{applyString}
\alias{applyString}
\title{Apply a function on elements of a character string}
\usage{
applyString(x, FUN, pos = NULL, pattern = NULL)
}
\arguments{
\item{x}{a character vector, or a vector to be coerced to a character vector.}

\item{FUN}{the function to be applied, see \code{\link[base:lapply]{base::lapply()}}.}

\item{pos}{a vector indicating the elements position.}

\item{pattern}{a pattern see \code{\link[base:gregexpr]{base::gregexpr()}}.}

\item{...}{argument to be passed to \code{\link[base:gregexpr]{base::gregexpr()}}.}
}
\value{
A character vector.
}
\description{
Apply a function on a given set of elements of a character string.
}
\note{
In case both \code{pos} or \code{pattern}, the latter is ignored.
}
\examples{
applyString('cool',  pos = 1:2, FUN = toupper)
applyString(c('cool', 'pro'),  pattern = 'o', FUN = toupper)
}
\author{
Kevin Cazelles
}
