% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignClass2df.R
\name{assignClass2df}
\alias{assignClass2df}
\title{Assign a class to a set of column of a data frame}
\usage{
assignClass2df(x, colid, cls)
}
\arguments{
\item{x}{a data frame or a R object to be coerced into a data frame.}

\item{colid}{the identity of columns for which class are ti be changed.}

\item{cls}{a character vector containing the classes' names to be used in the
same order as colid. By default \code{cls} is repeated until its size equals
\code{colid}'s size.}
}
\value{
A data.frame for which columns have the required classes.
}
\description{
This function is designed to ease the changes of column's class
of a given data frame.
}
\examples{
df1 <- matrix(signif(runif(20),4), ncol=2)
df2 <- assignClass2df(df1, 2, 'character')
str(df1)
str(df2)
}
