% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize.R
\name{categorize}
\alias{categorize}
\title{Assign categories to a vector of values.}
\usage{
categorize(x, categ, lower = FALSE)
}
\arguments{
\item{x}{A numeric, complex, character or logical vector.}

\item{categ}{A set of threshold that are used to assign categories.}

\item{lower}{A logical, if \code{TRUE} then \code{categ} are considered elements equal to a given threshold values are included in the lower category, default is FALSE.}
}
\value{
A vector of categories assinged.
}
\description{
Assigns a category to each element of a vector for a given set of threshold
values. Note that categories are always sorted. By defaults trheshold values
are regarded as upper boiundaries.
}
\examples{
categorize(stats::runif(40), categ=c(0.5,0.75))
categorize(LETTERS[1:5], categ='C')
categorize(LETTERS[1:5], categ='C', lower=TRUE)
categorize(LETTERS[floor(5*stats::runif(20))+1], categ=LETTERS[1:5], lower=TRUE)
}
\author{
Kevin Cazelles
}
