% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commaAnd.R
\name{commaAnd}
\alias{commaAnd}
\title{Paste element and add elememt separators.}
\usage{
commaAnd(x, comma = ", ", and = " and ")
}
\arguments{
\item{x}{vector of to be coerced to character strings (see \code{\link[=paste]{paste()}}).}

\item{comma}{element separator.}

\item{and}{last element separator (for 3 elements in \code{x} and more).}
}
\description{
Paste element and add elememt separators.
}
\examples{
commaAnd(c("Judith", "Peter", "Rebecca"))
}
