% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfTemplate.R
\name{dfTemplate}
\alias{dfTemplate}
\alias{dfTemplateMatch}
\title{Create a data frame from scratch or based on one or two data frames.}
\usage{
dfTemplate(cols, nrows = 1, col_classes = NULL, fill = NA)

dfTemplateMatch(x, y, yonly = FALSE, ...)
}
\arguments{
\item{cols}{either a number of column or a vector of character used as
columns names of the data frame to be returned.}

\item{nrows}{row number.}

\item{col_classes}{vector of column classes for the desired data frame.
By default, the class is determined by \code{fill}.}

\item{fill}{character or number used to fill out the columns. Default is \code{NA}.}

\item{x}{a data frame.}

\item{y}{data frame or a vector of strings that includes part of or all of
the column names of the data frame to be output.}

\item{yonly}{a logical. Should only \code{y} (or the \code{names(y)}) be used for the
data frame to be returned? Default is set to \code{FALSE} meaning that both the
names of \code{x} and the names of \code{y} are used.}

\item{...}{further arguments to be passed to \code{dfTemplate}.}
}
\value{
Returns a data frame with the desired characteristics.
}
\description{
This function handles the creation of data frames based on intuitive
parameters. It was originally designed to make row binding easier when
columns differs among data frame by creating data frames with the same
columns.
}
\section{Functions}{
\itemize{
\item \code{dfTemplateMatch}: Returns a data frame that matches names in \code{y}.
}}

\examples{
dfTemplate(5, 2)
dfTemplate(5, 2, col_classes = "character")
}
