% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicateRow.R
\name{duplicateRow}
\alias{duplicateCol}
\alias{duplicateRow}
\title{Duplicate element of a dataframe.}
\usage{
duplicateRow(x, id.el = 1, times = 1, append = FALSE)

duplicateCol(x, id.el = 1, times = 1, append = FALSE)
}
\arguments{
\item{x}{a data.frame.}

\item{id.el}{identity of the elements to be duplicated.}

\item{times}{number of times elements are duplicated. Could be a vector of the same length as id.el.}

\item{append}{A logical. If \code{TRUE}, duplicated elements will be appended to the dataframe.}
}
\description{
Duplicates rows and colmns of a given a dataframe.
}
\section{Functions}{
\itemize{
\item \code{duplicateRow}: A dataframe with duplicated rows.

\item \code{duplicateCol}: A dataframe with duplicated columns.
}}
\examples{
data(iris, package='datasets')
iris2 <- duplicateRow(iris, id.el=1:50, times=2)
iris3 <- duplicateCol(iris, id.el=c('Petal.Length', 'Petal.Width'), times=c(1,2), append=TRUE)
}
\author{
Kevin Cazelles
}

