% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findRef.R
\name{findRef}
\alias{findRef}
\title{Find citation tags.}
\usage{
findRef(text, con, markdown = FALSE)
}
\arguments{
\item{text}{A character string to be searched within.}

\item{con}{A connection object or a character string that stand for a path,
ignored if \code{text} is used.}

\item{markdown}{logical. If \code{TRUE}, citation tags of Markdown langage are sought.}
}
\value{
A data.frame with the citation keys and their frequency.
}
\description{
Finds Markdown and Latex citation tags in a given character stings or a file.
}
\examples{
test <- findRef(text='First \\\\cite{Pimm2000}, second \\\\Citep{May1972}')
}
