% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findReplace.R
\name{findReplace}
\alias{findReplace}
\alias{grepReplace}
\title{Find and replace}
\usage{
findReplace(vec, vec_pat, y)

grepReplace(vec, vec_pat, y)
}
\arguments{
\item{vec}{a vector to be searched.}

\item{vec_pat}{vector providing the patterns to be searched for.}

\item{y}{values to replace matching values.}
}
\description{
Functions to locate certain pattern and replace them.
}
\details{
These functions may be useful in case where many different patterns must be
replaced by the same value. Note that there are many other way to proceed.
}
\section{Functions}{
\itemize{
\item \code{findReplace}: Return a vector including the desire modifications.

\item \code{grepReplace}: Return a vector including the desire modifications.
}}

\examples{
vec <- c('fish', 'crab', 'crabs', 'bug', 'crab')
findReplace(vec, 'crab', 'fish')
findReplace(vec, c('crab', 'crabs'), 'fish')
grepReplace(vec, 'crab[s]?', 'fish')
}
