% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findThem.R
\name{findThem}
\alias{findThem}
\title{Find values in a given vector.}
\usage{
findThem(what, where, data = NULL)
}
\arguments{
\item{what}{a vector a values to be searched for.}

\item{where}{a vector where values will be searched on.}

\item{data}{an optional dataframe to be added to the search path.}
}
\value{
A two-columns dataframe indicating the position of found elements in simple data frame
}
\description{
Given a set of values and a vetor where the values must be found, the
\code{findThem} function records the matched values and the posiiton in the
vector where the values have been searched for.
}
\examples{
findThem(c(10,4,100), stats::rpois(1000,10))
}
\author{
Kevin Cazelles
}
\seealso{
\code{\link[base]{which}}
}

