% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianShape.R
\name{gaussianShape}
\alias{gaussianShape}
\title{Flexible bell-shaped function}
\usage{
gaussianShape(x, optx = 0, opty = 1/sqrt(2 * pi), width = 1, pow = 2)
}
\arguments{
\item{x}{a numeric vector.}

\item{optx}{x-values at which the maximum is reached.}

\item{opty}{extremum value.}

\item{width}{width of the bell.}

\item{pow}{a real number.}
}
\value{
A vector containing values standing for categories into which elements of x have fallen.
}
\description{
The \code{gaussianShape()} function computes a personnalisable Gaussian function.
Default values are the same as in \code{\link[stats:dnorm]{stats::dnorm()}}.
}
\examples{
gaussianShape(0)
plot(gaussianShape(1:1000, 500, 2, 250, pow=5), type='l')
}
\seealso{
\code{\link[stats:dnorm]{stats::dnorm()}}
}
\author{
Kevin Cazelles
}
