% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExtension.R
\name{getExtension}
\alias{getExtension}
\alias{getFilename}
\alias{getName}
\alias{getBasename}
\alias{getLocation}
\alias{getDetails}
\title{Extract file name, extension and basename ofrom a path.é}
\usage{
getExtension(x, sep = .Platform$file.sep)

getFilename(x, sep = .Platform$file.sep)

getName(x, sep = .Platform$file.sep)

getBasename(x, sep = .Platform$file.sep)

getLocation(x, sep = .Platform$file.sep)

getDetails(x, sep = .Platform$file.sep)
}
\arguments{
\item{x}{a character string. Note that \code{getDetails()} also supports vector of paths.}

\item{sep}{file separator, default used the platform-specific file separator see \code{\link[=.Platform]{.Platform()}}.}
}
\value{
Extract the file extension a character string (e.g. a path).
}
\description{
Extract file name, extension and basename ofrom a path.é
}
\details{
For more functionalities, have a look at package fs.
}
\section{Functions}{
\itemize{
\item \code{getFilename}: Extract the file name from a character string, that is the base name and the file extension

\item \code{getName}: Extract the name of a file or a directory from a character string.

\item \code{getBasename}: Extract the base name of a file from a character string.

\item \code{getLocation}: Extract the location of the parent folfder from a character string.

\item \code{getDetails}: Return a data frame with basic information for all elements of a vector of paths.
}}

\examples{
getExtension("path1/path2/foo.R")
getFilename("path1/path2/foo.R")
getBasename("path1/path2/foo.R")
getLocation("path1/path2/foo.R")
getDetails("path1/path2/foo.R")
getDetails(list.files(recursive = TRUE)) 
getExtension("foo.R")
getBasename("foo.R")
}
