% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inSilecoMisc.R
\docType{package}
\name{inSilecoMisc-package}
\alias{inSilecoMisc}
\alias{inSilecoMisc-package}
\title{\pkg{inSileco} is a set of miscellaneous R functions of inSileco.}
\description{
Mainly the functions of the packages used the manipulation of data
using base R, \emph{i.e.} outside the tidyverse which may (or may not)
has equivalent functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/inSileco/inSilecoMisc}
  \item \url{http://insileco.github.io/inSilecoMisc}
  \item Report bugs at \url{https://github.com/inSileco/inSilecoMisc/issues}
}

}
\author{
\strong{Maintainer}: Kevin Cazelles \email{kcazelle@uoguelph.ca} (\href{https://orcid.org/0000-0001-6619-9874}{ORCID})

Authors:
\itemize{
  \item David Beauchesne
  \item Nicolas Casajus
  \item Steve Vissault
}

}
\keyword{internal}
