% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keepWords.R
\name{keepWords}
\alias{keepWords}
\alias{keepLetters}
\alias{keepInitials}
\alias{wordCount}
\title{Keep a selection of words or letters}
\usage{
keepWords(
  str,
  slc = 1,
  collapse = " ",
  na.rm = FALSE,
  split_words = "[[:punct:][:space:]]+"
)

keepLetters(
  str,
  slc = 1,
  collapse = "",
  na.rm = FALSE,
  rm_punct = "[[:punct:][:space:]]+"
)

keepInitials(str, split_words = "[\\n\\t\\r\\f\\b[:punct:] ]+", collapse = "")

wordCount(str, split_words = "[[:punct:][:space:]]+")
}
\arguments{
\item{str}{an input character vector (or a list) from which words will be
extracted.}

\item{slc}{a vector of integer indicating the selected positions of the
words (or letters) to be kept.}

\item{collapse}{character string used to separate selected words (or
letters), if \code{NULL}, then selection is not collapsed and a list is returned.}

\item{na.rm}{a logical. Should missing values be removed?}

\item{split_words}{a character string containing a regular expression used
to split words.}

\item{rm_punct}{a character string containing a regular expression used to
remove punctuation characters.}
}
\value{
A vector (or a list) of the selected words.
}
\description{
Select words or letters based on their position in character strings.
}
\section{Functions}{
\itemize{
\item \code{keepLetters()}: A vector (or a list) of the selected letters.

\item \code{keepInitials()}: A vector (or a list) of initials.

\item \code{wordCount()}: A vector of the number of words for every character
strings passed as an input.

}}
\examples{
keepWords(loremIpsum(), 1:3)
keepWords(c(loremIpsum(),'Another character string!'), slc = c(1,4))
keepWords(c(loremIpsum(),'A second character string.'), slc = c(1,4),
 na.rm = TRUE, collapse = '/')
strex <- c('Lorem ipsum', 'dolor sit', ' amet;')
keepLetters(strex, c(1,4))
keepLetters(strex, c(1,4), collapse = "")
keepInitials("National Basketball Association")
wordCount(c("two words!", "... and three words"))
wordCount(loremIpsum())
}
\seealso{
\code{\link[=strsplit]{strsplit()}}
}
