% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{logistic}
\alias{logistic}
\alias{logistic2}
\title{Logistic functions}
\source{
\href{https://en.wikipedia.org/wiki/Logistic_function}{wikipedia.org/wiki/Logistic_function}
}
\usage{
logistic(x, yneg = -1, ypos = 1, lambda = 1, pow = 1)

logistic2(x, yneg = -1, ypos = 1, lambda = 1, pow = 1, yzer = 0)
}
\arguments{
\item{x}{a numerical vector.}

\item{yneg}{assympotic values when x tends to \code{-Inf}.}

\item{ypos}{assympotic values when x tends to \code{-Inf}.}

\item{lambda}{scalar coefficient.}

\item{pow}{x exponent.}

\item{yzer}{values (for \code{logistic2} only).}
}
\value{
A numeric vector.
}
\description{
The \code{logisic} function describe the classical logistic function,
}
\details{
The classic logistic equation is:

\deqn{f(x) = \frac{ypos-yneg}{1+e^{-\lambda x^{pow}}}}{%
      f(x) = (ypos-yneg)/(1+exp(-\lambda x^pow))}

A slighly different version is:

\deqn{f(x) = yneg + \frac{1}{\frac{1}{ypos-yneg}+(\frac{1}{yzer-yneg}-\frac{1}{ypos-yneg})e^{-\lambda x^{pow}}}}{%
      f(x) = yneg + 1/(1/(ypos-yneg)+(1/(yzer-yneg)-1/ypos-yneg)exp(-\lambda x^pow))}
}
\section{Functions}{
\itemize{
\item \code{logistic2}: a slighlty different logistic function..
}}

