% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{logit}
\alias{logit}
\title{Logit and its inverse.}
\usage{
logit(x, inv = FALSE)
}
\arguments{
\item{x}{a object to be coerced to numeric.}

\item{inverse}{a logical. Should the inverse logit be computed?}
}
\description{
Compute the logit function and its inverse. Note that package \code{boot}
provides \code{logit} and \code{inv.logit} functions.
}
\examples{
logit(-10:10, inv=TRUE)
}
\author{
Kevin Cazelles
}
