% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loremIpsum.R
\name{loremIpsum}
\alias{loremIpsum}
\title{Lorem ipsum.}
\usage{
loremIpsum(n = NULL)
}
\arguments{
\item{n}{the number of words to be kept.}
}
\value{
A character string with the desired piece of lorem ipsum.
}
\description{
Returns a common form of the Lorem ipsum text an optionally a subset of it.
}
\examples{
loremIpsum(5)
}
\references{
https://en.wikipedia.org/wiki/Lorem_ipsum
}
