% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiMatch.R
\name{multiMatch}
\alias{multiMatch}
\title{Multiple match}
\usage{
multiMatch(text, pattern)
}
\arguments{
\item{text}{a character vector where matches are sought, or an object which
can be coerced by as.character to a character vector, see \code{regexec}.}

\item{pattern}{character string containing a regular expression to be
matched in the given character vector, see \code{regexec}.}
}
\value{
A vector of character strings matching \code{pattern} argument.
}
\description{
Seek all elements matching a given pattern in a character strings.
}
\details{
\code{multiMatch} is essentially a wrapper around \code{\link[base:grep]{base::regexec()}} and
\code{\link[base:regmatches]{base::regmatches()}} to find more than one match.
}
\examples{
multiMatch(loremIpsum(), " [[:alnum:]]{3} ")
}
