% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqRg.R
\name{seqRg}
\alias{seqRg}
\title{Generate a regular sequence using the range of a vector}
\usage{
seqRg(x, n, offset = 0, prop = TRUE)
}
\arguments{
\item{x}{a vector.}

\item{n}{the number of values in the output sequence.}

\item{offset}{extend or reduce the range .}

\item{prop}{a logical. If \code{TRUE}, then \code{offset} is a proportion of the range, otherwise \code{offset} is used as is.}
}
\description{
Generate a regular sequence using the range of a vector
}
\examples{
seqRg(runif(10), 10)
seqRg(1:10, 10, .5)
}
