% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setColClass.R
\name{setColClass}
\alias{setColClass}
\title{Set class of data frames columns}
\usage{
setColClass(x, colid, cls)
}
\arguments{
\item{x}{a data frame or an R object to be coerced into a data frame.}

\item{colid}{the identity of columns for which class are to be changed.}

\item{cls}{a character vector containing the classes' names to be used in the
same order as \code{colid}. By default, \code{cls} is repeated until its size equals
\code{colid}'s size.}
}
\value{
A data frame whose columns have the desired classes.
}
\description{
This function is designed to ease the changes of column's class
of a given data frame.
}
\examples{
df1 <- matrix(signif(runif(20),4), ncol=2)
df2 <- setColClass(df1, 2, 'character')
str(df1)
str(df2)
}
