% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signifSymbols.R
\name{signifSymbols}
\alias{signifSymbols}
\title{A simple function to associated p-values with symbols.}
\usage{
signifSymbols(pvalue, thresholds = c(0.1, 0.05, 0.01, 0.001),
  symbols = c(".", "*", "**", "***"), notsignif = "n.s.")
}
\arguments{
\item{pvalue}{a pvalue for which a symbol is requested.}

\item{thresholds}{the threshold values that define category to which symbols are assigned.}

\item{symbols}{list of symbols.}

\item{notsignif}{symbols for non significant p-value.}
}
\description{
\code{signifSymbols} take one p-value and associated to symbols. The default
values used the most common symbols.
}
\examples{
signifSymbols(.012)
signifSymbols(.008)
sapply(c(.2, .08, .04, .008, 0.0001), signifSymbols)
}
\author{
Kevin Cazelles
}
