% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squaretize.R
\name{squaretize}
\alias{squaretize}
\title{Makes a data frame or a matrix square}
\usage{
squaretize(x, fill = 0, reorder = TRUE)
}
\arguments{
\item{x}{an R object to be coerced into matrix.}

\item{fill}{element used to fill the additional rows and/or columns.}

\item{reorder}{a logical indicating whether the names must be used to order rows and columns.}
}
\value{
A square matrix.
}
\description{
Add missing rows or columns based on names and Duplicates rows and columns
of a given a data frame.
}
\examples{
mat <- matrix(1:12, 3, 4)
mat2 <- squaretize(mat)
matb <- matrix(1:12, 4, 3)
colnames(matb) <- LETTERS[1:3]
mat2b <- squaretize(matb)
}
