% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stLength.R
\name{stLength}
\alias{stLength}
\title{Compute the length of a strings.}
\usage{
stLength(s, ignore = "")
}
\arguments{
\item{s}{character vector or a object to be coerced as a character vector.}

\item{ignore}{a vector of characters to be ignored in the count.}
}
\description{
Compute the length of a vector of strings.
}
\examples{
stLength(c(123))
stLength(c('/', 'four five'), ignore = c(" ", "f"))
}
\author{
Kevin Cazelles
}
