% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwatch.R
\name{stopwatch}
\alias{stopwatch}
\alias{timer}
\title{Timer and Stopwatch}
\usage{
stopwatch(max_time = 1000, pause = 0.01, digits = 5)

timer(time, pause = 0.01, digits = 5)
}
\arguments{
\item{max_time}{time in seconds after which the stopwatch is forced to stop, use \code{Inf} to run it endlessly.}

\item{pause}{time duration of the pause between 2 time increments displayed (it should better be small).}

\item{digits}{number of digits displayed.}

\item{time}{initial time (in seconds) for the timer.}
}
\description{
Timer and Stopwatch
}
\section{Functions}{
\itemize{
\item \code{timer}: Timer
}}

\examples{
\dontrun{
stopwatch(10)
stopwatch(20)
}
}
