% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substrBib.R
\name{substrBib}
\alias{substrBib}
\title{Substract Bibtex entries.}
\usage{
substrBib(bib, text, con, markdown = FALSE)
}
\arguments{
\item{bib}{An object of class \code{BibEntry} or \code{bibentry}.}

\item{text}{A character string to be searched within.}

\item{con}{A connection object or a character string that stand for a path,
ignored if text is defined.}

\item{markdown}{logical. If \code{TRUE}, Markdown formated citation marks are sought.}
}
\description{
Substarct Bibtex entries that are encontered in a given text.
}
\examples{
file.name <- system.file('Bib', 'RJC.bib', package='RefManageR')
bib <- RefManageR::ReadBib(file.name)
# bib[[200]]
txt <- 'First \\\\cite{kim1995estimating}, second \\\\Citep{fu2006statistical}'
out <- substrBib(bib, txt)
# RefManageR::WriteBib(out, file=tempfile(fileext = '.bib'))
}
\author{
Kevin Cazelles
}
