% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tblDown.R
\name{tblDown}
\alias{tblDown}
\title{Write data frames in a document}
\usage{
tblDown(
  x,
  output_file = "./tables.docx",
  section = NULL,
  caption = NULL,
  title = NULL,
  row.names = FALSE,
  ...
)
}
\arguments{
\item{x}{a data frame or a list of data frames.}

\item{output_file}{path to the output file. Its extension will be used by
pandoc to correctly render the final document in the write format.}

\item{section}{a vector of character strings used as section titles (optional).}

\item{caption}{a vector of character strings used as captions (optional).}

\item{title}{a character string used as a title for the document (optional).}

\item{row.names}{a logical. Should row names be added? See \code{\link[knitr:kable]{knitr::kable()}}.}

\item{...}{further arguments passed to \code{\link[knitr:kable]{knitr::kable()}}.}
}
\value{
A data frame whose columns have the desired classes.
}
\description{
Write a data frame or a list of data frames in a markdown document converted #' in various format using pandoc.
}
\details{
This function calls \code{\link[base:cat]{base::cat()}} and \code{\link[knitr:kable]{knitr::kable()}} to write a Markdown
document containing a list of tables that is then converted into the
desired format. For \code{section} and \code{caption} if the length differ then
will be cut off or expanded.
}
\examples{
\dontrun{
data(CO2)
tblDown(list(CO2[1:2, ], CO2[3:6,]), section = "section")
tblDown(list(CO2[1:2, ], CO2[3:6,]), "./tables.pdf")
}
}
\references{
\url{https://pandoc.org/MANUAL.html#tables}
}
