% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whichIs.R
\name{whichIs}
\alias{whichIs}
\title{Search values or patterns in a vector}
\usage{
whichIs(x, y, isPattern = FALSE, ...)
}
\arguments{
\item{x}{a vector of values or patterns to be searched for.}

\item{y}{a vector where x values are searched for.}

\item{isPattern}{should x be considered as a vector of patterns?}

\item{...}{further arguments to be passed to \code{\link[=which]{which()}}.}
}
\value{
A list of positions. Each element corresponds to the matches for a specific x value.
}
\description{
Search values or patterns in a vector.
}
\examples{
vec <- LETTERS[1:10]
spl <- sample(vec)
id <- unlist(whichIs(vec, spl))
identical(vec, spl[id])
}
