% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordCount.R
\name{wordCount}
\alias{wordCount}
\title{Count words}
\usage{
wordCount(string, exclude = "[[:punct:]]")
}
\arguments{
\item{string}{Input vector. A character vector or list from which words will be extracted.}

\item{exclude}{character string containing a regular expression listing the character to be excluded.}
}
\value{
A vector of the number of word in each element of \code{sting}.
}
\description{
Function that counts the number of word in a character string.
}
\examples{
strex <- 'Lorem ipsum dolor sit amet, consectetur adipisicing elit.'
wordCount(strex)
wordCount(c(strex,'A second character string.'))
}
\author{
Kevin Cazelles
}
