% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addURL.R
\name{addURL}
\alias{addURL}
\title{Computes the}
\usage{
addURL(url, text = NULL, extra = NULL, markdown = FALSE)
}
\arguments{
\item{url}{An URL.}

\item{text}{A character string to be searched within.}

\item{extra}{Extra A connection object or a character string that stand for a path,
ignored if text is defined.}

\item{markdown}{logical. If \code{TRUE}, hyperlinks  marks of Markdown language are sought.}
}
\description{
Substarct Bibtex entries that are encontered in a given text.
}
\examples{
addURL('https://github.com/inSileco/letiRmisc', 'Github repo')
addURL('https://github.com/inSileco/letiRmisc', 'Github repo', markdown=TRUE)
}
\author{
Kevin Cazelles
}
