% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateCol.R
\name{aggregateCol}
\alias{aggregateCol}
\title{Aggregate columns given an integer vector.}
\usage{
aggregateCol(data, grp, names_aggreg = NULL, FUN = sum, ...)
}
\arguments{
\item{data}{a dataframe.}

\item{grp}{an integer vector assigning a group to each column of \code{data}
used to compute operations on a subset of columns. Any integer is a valid group,
\code{0} indicates that the column must be unchanged, \code{NA} is used to remove a
column.}

\item{names_aggreg}{column names for aggregated columns.}

\item{FUN}{a function to be applied to all group of columns.}

\item{...}{further arguments to be passed to \code{FUN}.}
}
\value{
A dataframe with the grouped columns.
}
\description{
The \code{aggregateCol} makes the computation of a given function on
subsets of columns easy. An integer vector, whose length is the number of
column of the input dataframe, is used to assign a group to each column
that explicits the subsetting. Furthermore columns can be either kept as is
using \code{0} or discarded by using \code{NA}.
}
\examples{
mat1 <- matrix(1:70,10)
grp1 <- c(NA,0,rep(1,5))
aggregateCol(mat1, grp1, FUN = mean)
}
\seealso{
\code{\link[stats]{aggregate}}
}
\author{
Kevin Cazelles
}
