% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignIds.R
\name{assignIds}
\alias{assignIds}
\title{Assign an id to a list of characters or factors}
\usage{
assignIds(vec, alphabetical = FALSE)
}
\arguments{
\item{alphabetical}{a logical indicating whether an alphabetical sorting must be applied.}

\item{x}{object to be coerced into character type}
}
\description{
This function is done mainly to assign an id to factor or similar character
strings. Regarding factors, \code{as.integer} does such task
but the order obtained could be different.
}
\examples{
assignIds(list(2,'f', 'd', 'f'))
}

