% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize.R
\name{categorize}
\alias{categorize}
\title{Assign categories to a vector of values.}
\usage{
categorize(x, categ, lower = FALSE, sort_categ = FALSE)
}
\arguments{
\item{x}{A numeric, complex, character or logical vector.}

\item{categ}{A set of threshold that are used to assign categories.}

\item{lower}{A logical, if TRUE elements equal to a given threshold values are included in the lower category, default is FALSE.}

\item{sort_categ}{A logical indicating whether categories should be sorted.}
}
\value{
A vector countaining values standing for categories into which elements of x have fallen.
}
\description{
Assigns a category to each element of a vector for a given set of threshold values.
}
\examples{
categorize(stats::runif(40), categ=c(0.5,0.75))
categorize(LETTERS[1:5], categ='C')
categorize(LETTERS[1:5], categ='C', lower=TRUE)
categorize(LETTERS[floor(5*stats::runif(20))+1], categ=LETTERS[1:5], lower=TRUE)
}
\author{
Kevin Cazelles
}

