% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize.R
\name{categorize}
\alias{categorize}
\title{Assign categories to a vector of values.}
\usage{
categorize(x, categ, lower = FALSE)
}
\arguments{
\item{x}{A numeric, complex, character or logical vector.}

\item{categ}{A set of threshold values used to assign categories.}

\item{lower}{A logical, if \code{TRUE} threshold values (i.e. values within
\code{categ}) belongs the the lower category rather than the upper (default
behaviour).}
}
\value{
A vector of categories assigned.
}
\description{
Assigns a category to each element of a vector for a given set of threshold
values..
}
\examples{
categorize(stats::runif(40), categ=c(0.5,0.75))
categorize(LETTERS[1:5], categ='C')
categorize(LETTERS[1:5], categ='C',  lower=TRUE)
categorize(LETTERS[floor(5*stats::runif(20))+1], categ=LETTERS[1:5], lower=TRUE)
}
\author{
Kevin Cazelles
}
