% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicateRow.R
\name{duplicateRow}
\alias{duplicateRow}
\alias{duplicateCol}
\title{Duplicates elements of a data frame}
\usage{
duplicateRow(x, id.el = 1, times = 1, append = FALSE)

duplicateCol(x, id.el = 1, times = 1, append = FALSE)
}
\arguments{
\item{x}{a data frame.}

\item{id.el}{identity of the elements to be duplicated.}

\item{times}{number of times elements are duplicated. Could be a vector of the same length as \code{id.el}.}

\item{append}{A logical. If \code{TRUE}, duplicated elements will be appended to the data frame otherwise duplicated elements remain next to their parent. Non-existing columns cannot be duplicated while non-existing rows can and produce \code{NA}.}
}
\value{
returns a data frame with duplicated rows.
}
\description{
Duplicates rows and columns of a given a data frame.
}
\section{Functions}{
\itemize{
\item \code{duplicateCol()}: returns a data frame with duplicated columns.

}}
\examples{
data(iris, package = "datasets")
iris2 <- duplicateRow(iris, id.el = 1:50, times = 2)
iris3 <- duplicateCol(
iris,
id.el = c("Petal.Length", "Petal.Width"),
times = c(1, 2),
append = TRUE)

}
