% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findThem.R
\name{findThem}
\alias{findThem}
\title{Find values in a given vector.}
\usage{
findThem(what, where, data = NULL, todf = FALSE)
}
\arguments{
\item{what}{a vector a values to be searched for.}

\item{where}{a vector where values will be searched on.}

\item{data}{an optional dataframe to be added to the search path.}

\item{todf}{a logical indicating whether the output must be a dataframe.}
}
\value{
A list indicating matched positions for each elements of \code{what}. If \code{todf} is
TRUE then a three-columns dataframe is returned includeing values and positions
in both \code{what} and \code{where} vectors.
}
\description{
Given a set of values and a vetor where the values must be found, the
\code{findThem} function records the matched values and the posiiton in the
vector where the values have been searched for.
}
\examples{
wha <- stats::rpois(1000,10)
findThem(c(10,4,100), wha)
findThem(c(10,4,100), wha, todf=TRUE)
}
\author{
Kevin Cazelles
}
\seealso{
\code{\link[base]{which}}
}

