% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findThem.R
\name{findThem}
\alias{findThem}
\title{Find values in a given vector.}
\usage{
findThem(what, where, todf = FALSE, reportnomatch = FALSE)
}
\arguments{
\item{what}{a vector a values to be searched for.}

\item{where}{a vector where values will be searched on.}

\item{todf}{a logical indicating whether the output object must be a data frame.}

\item{reportnomatch}{a logical. If \code{TRUE}, values without match are reported #' in the data frame with a NA. Only available if \code{todf} is \code{TRUE}. Default is
set to \code{FALSE}.}
}
\value{
A list indicating matched positions for each elements of \code{what}. If no
match is found then \code{NA} is returned. If \code{todf} is \code{TRUE}
then a three-columns data frame is returned including values and positions
in both \code{what} and \code{where} vectors.
}
\description{
Given a set of values and a vector where the values must be found, the
\code{findThem} function records the matched values and the position in the
vector where the values have been searched for.
}
\examples{
x <- stats::rpois(1000,10)
findThem(c(10,4,100), x)
findThem(c(10,4,100), x, todf=TRUE)
findThem(c(10,4,100), x, todf=TRUE, reportnomatch=TRUE)
}
\seealso{
\code{\link[=which]{which()}}
}
