% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianShape.R
\name{gaussianShape}
\alias{gaussianShape}
\title{Flexible bell-shaped function.}
\usage{
gaussianShape(x, optx = 0, opty = 1/sqrt(2 * pi), width = 1, pow = 2)
}
\arguments{
\item{x}{a numeric vector.}

\item{optx}{the x values at which the maximum is reached.}

\item{opty}{the maximum values.}

\item{width}{the width of the bell-shaoed}

\item{pow}{a real number.}
}
\value{
a vector countaining values standing for categories into which elements of x have fallen.
}
\description{
The \code{gaussianShape} function computes a gaussian function readily customizable.
Default values are the same as in \code{\link[stats]{dnorm}}.
}
\examples{
gaussianShape(seq(-10,10,0.1))
plot(gaussianShape(1:1000, 500, 2, 250, pow=5), type='l')
}
\author{
Kevin Cazelles
}
\seealso{
\code{\link[stats]{dnorm}}
}

