% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDigits.R
\name{getDigits}
\alias{getDigits}
\title{Extract digits from a character vector.}
\usage{
getDigits(x, collapse = TRUE)
}
\arguments{
\item{x}{a character vector where digits are sought, or an object which can be coerced by \code{as.character} to a character vector.}

\item{collapse}{super useful.}
}
\value{
A list of digits.
}
\description{
This function extracts all digits found in character vector and returns them
as a list.
}
\details{
This function is very useful!
}
\examples{
getDigits('cool012-99')
}
\author{
Kevin Cazelles
}
