% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keepWords.R
\name{keepWords}
\alias{keepWords}
\alias{keepLetters}
\alias{rmPunct}
\title{Keep words and letters.}
\usage{
keepWords(str, slc = 1, punct.rm = TRUE, na.rm = FALSE,
  collapse = NULL, unlist = FALSE)

keepLetters(str, slc = 1, punct.rm = FALSE, unlist = TRUE)

rmPunct(str, replacement = "")
}
\arguments{
\item{str}{An input character vector or a list from which words will be
extracted.}

\item{slc}{Vector of integer indicating the selected positions of the words to be kept.}

\item{punct.rm}{A logical Should punctuation characters be removed? Note that this affects the order of character in a string.}

\item{na.rm}{A logical. Should missing values be removed?}

\item{collapse}{An optional character string used to separate selected words.}

\item{unlist}{A logical. Should the results be unlisted?}

\item{replacement}{a replacement for matched pattern in \code{\link[base:gsub]{base::gsub()}}.}
}
\value{
A vector of the selected words.
}
\description{
Extract words and letters based on their position in character strings, vectorised over the input character string vector.
}
\section{Functions}{
\itemize{
\item \code{keepLetters}: A vector containing the selection of letters.

\item \code{rmPunct}: remove punctuation
}}

\examples{
strex <- 'Lorem ipsum dolor sit amet, consectetur adipisicing elit.'
keepWords(strex)
keepWords(c(strex,'A second chacter string.'), slc=c(1,8), na.rm=TRUE)
keepWords(c(strex,'A second chacter string.'), slc=c(1,8), na.rm=TRUE, collapse='/')
strex <- c('Lorem ipsum', 'dolor sit', ' amet;')
keepLetters(strex, c(1,4))
}
\author{
Kevin Cazelles
}
