% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keepWords.R
\name{keepWords}
\alias{keepWords}
\title{Keep these words.}
\usage{
keepWords(string, slc = 1, punct.rm = TRUE, na.rm = FALSE,
  collapse = NULL)
}
\arguments{
\item{string}{character Input vector. A character vector or list from which words will be extracted.}

\item{slc}{Vector of integer indicating the selected positions of the words to be kept.}

\item{punct.rm}{A logical indicating whether punctutation characters should be removed.}

\item{na.rm}{A logical indicating whether missing values should be removed.}

\item{collapse}{An optional character string used to separate selected words. Not \code{NA_character_}.}
}
\value{
A vector of the selected words concatened for each string.
}
\description{
Extract words based on their position in a character string, vectorised over \code{string}.
}
\examples{
strex <- 'Lorem ipsum dolor sit amet, consectetur adipisicing elit.'
keepWord(strex)
keepWord(c(strex,'A second chacter string.'), slc=c(1,8), na.rm=TRUE)
keepWord(c(strex,'A second chacter string.'), slc=c(1,8), na.rm=TRUE, collapse='/')
}
\author{
Kevin Cazelles
}

