% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanAlong.R
\name{meanAlong}
\alias{meanAlong}
\title{Compute the mean along a vector}
\usage{
meanAlong(vec, n)
}
\arguments{
\item{vec}{a vector of numeric}

\item{n}{a intger indicating the size of the window}
}
\description{
This function is a simple moving window function.
}
