% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgInfo.R
\name{msgInfo}
\alias{msgInfo}
\alias{msgError}
\alias{msgSuccess}
\alias{msgWarning}
\title{Pre-formatted message functions}
\usage{
msgInfo(...)

msgError(...)

msgSuccess(...)

msgWarning(...)
}
\arguments{
\item{...}{text to be passed to \code{\link[=paste]{paste()}}.}
}
\value{
Reports an info.
}
\description{
Convenient functions to sent notices to the user while a function/script is
being executed.
}
\section{Functions}{
\itemize{
\item \code{msgError}: Reports an Error.

\item \code{msgSuccess}: Reports a success.

\item \code{msgWarning}: Reports a warning.
}}

\note{
All of these functions call \code{\link[=message]{message()}}, so for any function \code{FUN()} using #' them, \code{suppressMessages(FUN())} is sufficient to mute all messages.
}
\examples{
msgInfo("computing")
msgSuccess("done")
msgError("you got it wrong")
msgWarning("be careful")
}
\seealso{
\link[cli:symbol]{cli::symbol} \code{\link[=message]{message()}}
}
