% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleWithin.R
\name{scaleWithin}
\alias{scaleWithin}
\title{Scale a set of values}
\usage{
scaleWithin(x, nval = 100, mn = min(x), mx = max(x))
}
\arguments{
\item{x}{a numeric object of type \code{numeric} or coercible as one.}

\item{nval}{an integer designing The number of scaled values (between 1 and \code{nval}).}

\item{mn}{minimum value (\code{mn} and lower values are set to 1).}

\item{mx}{maximum value (\code{mx} and higher values are set to \code{nval}).}
}
\value{
A numeric vector
}
\description{
This function maps a set of values to a range of integers (from 1
to a positive integer passed as an argument).
}
\examples{
x <- stats::rpois(20, 20)
scaleWithin(x, 20, 10, 30)
scaleWithin(matrix(x, 10, 2), mn = 20, mx = 30)
}
