% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqRg.R
\name{seqRg}
\alias{seqRg}
\alias{seqCol}
\alias{seqRow}
\title{Generate a regular sequence based on the range of a vector}
\usage{
seqRg(x, n, offset = 0, prop = TRUE)

seqCol(df)

seqRow(df)
}
\arguments{
\item{x}{a vector.}

\item{n}{the number of values in the output sequence.}

\item{offset}{extend or reduce the range .}

\item{prop}{a logical. If \code{TRUE}, then \code{offset} is a proportion of the
range, otherwise \code{offset} is used as is.}

\item{df}{a data frame or a matrix.}
}
\description{
Generate a regular sequence based on the range of a vector
}
\section{Functions}{
\itemize{
\item \code{seqCol()}: a regular sequence based on the number of columns in a data frame.

\item \code{seqRow()}: a regular sequence based on the number of rows in a data frame.

}}
\examples{
seqRg(runif(10), 10)
seqRg(1:10, 10, .5)
seqCol(matrix(1, 3, 2))
}
